requirements.txt
# Core dependencies for CivitAI Flux Dev LoRA Tagging Assistant
Pillow>=9.0.0  # For image validation and processing
fastapi>=0.95.0  # Web framework with WebSocket support
uvicorn>=0.22.0  # ASGI server for running FastAPI
python-multipart>=0.0.6  # For handling file uploads
websockets>=11.0.3  # WebSocket protocol implementation

# Testing dependencies
pytest>=7.0.0  # Testing framework
httpx>=0.23.0  # HTTP client for testing FastAPI
pytest-asyncio>=0.19.0  # Async support for pytest
selenium>=4.1.0  # For browser compatibility testing (optional)
